requirements.txt
openai>=2.29.0
anthropic>=0.86.0
google-generativeai>=0.8.6
